requirements.txt
python-telegram-bot==22.6
SQLAlchemy==2.0.36
asyncpg==0.31.0
python-dotenv==1.0.1
cachetools==5.3.2
google-genai==1.0.0
youtube-transcript-api==0.6.3
httpx==0.27.2
beautifulsoup4==4.12.3
groq==0.13.1
pydub==0.25.1
PyMuPDF==1.25.3